feat: change group package
```
com.sgyj -> com.gg
```
