+++ build.gradle
@@ -11,7 +11,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.0'
 }
 
-group = 'com.sgyj'
+group = 'com.gg'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '17'
 
@@ -40,7 +40,7 @@ ext {
 }
 
 dependencies {
-    implementation 'com.sgyj:common-service:1.0.8'
+    implementation 'com.gg:common-service:1.0.0'
     implementation 'org.springframework.cloud:spring-cloud-starter-config'
     implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
     implementation 'org.springframework.cloud:spring-cloud-starter-bus-amqp'
@@ -51,6 +51,7 @@ dependencies {
 
     implementation 'org.springframework.cloud:spring-cloud-starter-sleuth'
     implementation 'org.springframework.cloud:spring-cloud-starter-zipkin:2.2.8.RELEASE'
+    implementation 'org.springframework.boot:spring-boot-starter-actuator'
     implementation 'io.micrometer:micrometer-registry-prometheus'
     implementation 'javax.ws.rs:javax.ws.rs-api:2.1.1'
     implementation 'org.glassfish.jersey.core:jersey-common:3.1.1'
