+++ src/main/java/com/gg/accountservice/modules/account/dto/kafka/Fields.java
@@ -1,0 +1,14 @@
+package com.gg.accountservice.modules.account.dto.kafka;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Fields {
+
+    private String type;
+    private boolean optional;
+    private String field;
+
+}
