+++ src/main/java/com/gg/accountservice/modules/account/dto/kafka/Payload.java
@@ -1,0 +1,11 @@
+package com.gg.accountservice.modules.account.dto.kafka;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class Payload {
+
+
+}
