+++ src/main/java/com/gg/accountservice/modules/account/form/AuthCodeForm.java
@@ -1,0 +1,18 @@
+package com.gg.accountservice.modules.account.form;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+@Data
+@RequiredArgsConstructor
+public class AuthCodeForm {
+
+    @NotNull(message = "이메일을 입력해 주세요.")
+    @Email(message = "이메일 형식이 올바르지 않습니다.")
+    private String email;
+
+    private String authCode;
+
+}
