+++ src/main/java/com/gg/accountservice/modules/account/service/AccountService.java
@@ -1,0 +1,146 @@
+package com.gg.accountservice.modules.account.service;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.gg.accountservice.modules.account.dto.CustomAccountDto;
+import com.gg.accountservice.modules.account.dto.TokenDto;
+import com.gg.accountservice.modules.account.entity.Account;
+import com.gg.accountservice.modules.account.form.AccountSaveForm;
+import com.gg.accountservice.modules.account.form.AuthCodeForm;
+import com.gg.accountservice.modules.account.repository.AccountRepository;
+import com.gg.accountservice.modules.account.service.kafka.KafkaEmailProducer;
+import com.sgyj.commonservice.advice.exceptions.ExpiredTokenException;
+import com.sgyj.commonservice.advice.exceptions.NoMemberException;
+import com.sgyj.commonservice.advice.exceptions.NotFoundException;
+import com.sgyj.commonservice.annotation.BaseServiceAnnotation;
+import com.sgyj.commonservice.dto.account.AccountDto;
+import com.sgyj.commonservice.enums.LoginType;
+import com.sgyj.commonservice.properties.KafkaUserTopicProperties;
+import com.sgyj.commonservice.security.CredentialInfo;
+import com.sgyj.commonservice.security.Jwt;
+import com.sgyj.commonservice.security.Jwt.Claims;
+import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@BaseServiceAnnotation
+@RequiredArgsConstructor
+public class AccountService {
+
+    private final PasswordEncoder passwordEncoder;
+    private final AccountRepository accountRepository;
+    private final Jwt jwt;
+    private final KafkaEmailProducer kafkaEmailProducer;
+    private final KafkaUserTopicProperties kafkaUserTopicProperties;
+
+    /**
+     * 2
+     * 회원가입
+     *
+     * @param accountSaveForm account 저장 폼
+     * @return AccountDto account 생성 결과 Dto
+     */
+    public AccountDto saveAccount(AccountSaveForm accountSaveForm) throws JsonProcessingException {
+        accountSaveForm.setPassword(passwordEncoder.encode(accountSaveForm.getPassword()));
+
+        validateAccount(accountSaveForm);
+
+        String authCode = sendSignUpConfirmEmail(accountSaveForm.getEmail());
+        Account account = Account.createAccountByFormAndAuthCode(accountSaveForm, authCode);
+        accountRepository.save(account);
+        // kafkaAccountProducer.send(kafkaUserTopicProperties.getAccountTopic(), accountDto);
+        return CustomAccountDto.from(account);
+    }
+
+    /**
+     * 이메일 인증코드 전송
+     *
+     * @param email 이메일
+     * @return authCode 인증코드
+     */
+    private String sendSignUpConfirmEmail(String email) throws JsonProcessingException {
+        String authCode = RandomStringUtils.randomAlphanumeric(12);
+        // kafkaEmailProducer.send(kafkaUserTopicProperties.getAuthenticationMailTopic(), new EmailMessage(email, "제목", authCode));
+        return authCode;
+    }
+
+    /**
+     * 입력된 정보 확인
+     *
+     * @param accountSaveForm account 저장 폼
+     */
+    private void validateAccount(AccountSaveForm accountSaveForm) {
+
+        if (validNickname(accountSaveForm.getNickname())) {
+            throw new BadCredentialsException("이미 존재하는 닉네임입니다.");
+        }
+
+        accountRepository.findByEmail(accountSaveForm.getEmail()).ifPresent(account -> {
+            throw new BadCredentialsException("이미 존재하는 이메일입니다.");
+        });
+
+    }
+
+    /**
+     * nickName 중복확인
+     *
+     * @param nickName 닉네임
+     * @return boolean 닉네임 유효값 확인 결과
+     */
+    public boolean validNickname(String nickName) {
+        return accountRepository.findByNickname(nickName).isPresent();
+    }
+
+    /**
+     * 사용자 로그인
+     *
+     * @param email      이메일
+     * @param credential 인증
+     * @return AccountDto 계정 Dto
+     */
+    public AccountDto login(String email, CredentialInfo credential) {
+        Account account = accountRepository.findByEmailAndLoginType(email, credential.getLoginType()).orElseThrow(() -> new NotFoundException("등록된 계정이 없습니다."));
+        account.login(passwordEncoder, credential.getCredential());
+        account.afterLoginSuccess();
+        return CustomAccountDto.createByAccountAndGenerateAccessToken(account, jwt);
+    }
+
+    /**
+     * 이메일 인증 확인
+     *
+     * @param authCodeForm 인증 코드 확인 form
+     * @return AccountDto 인증 확인 AccountDto
+     */
+    public AccountDto validAuthCode(AuthCodeForm authCodeForm) {
+        Account account = accountRepository.findByEmailAndLoginType(authCodeForm.getEmail(), LoginType.TGAHTER)
+            .orElseThrow(() -> new NotFoundException("등록된 계정이 없습니다."));
+
+        if (!account.getAuthCode().equals(authCodeForm.getAuthCode())) {
+            throw new BadCredentialsException("인증 코드가 잘못되었습니다. 다시 확인해주세요.");
+        }
+        account.successAuthUser();
+        return CustomAccountDto.createByAccountAndGenerateAccessToken(account, jwt);
+    }
+
+    /**
+     * 토큰 갱신
+     *
+     * @param tokenDto 토큰 갱신 Dto
+     * @return TokenDto 갱신 토큰 결과 Dto
+     */
+    public TokenDto renewalTokenByRefreshToken(TokenDto tokenDto) {
+        if (jwt.validateToken(tokenDto.getRefreshToken())) {
+            Claims claims = jwt.verify(tokenDto.getRefreshToken());
+            Account account = accountRepository.findByEmailAndLoginType(claims.getEmail(), LoginType.TGAHTER)
+                .orElseThrow(() -> new NotFoundException("이메일을 찾을 수 없습니다."));
+            AccountDto accountDto = CustomAccountDto.createByAccountAndGenerateAccessToken(account, jwt);
+            return TokenDto.builder().accessToken(accountDto.getAccessToken()).refreshToken(accountDto.getRefreshToken()).build();
+        }
+        throw new ExpiredTokenException();
+    }
+
+    public AccountDto getAccount(String accountId) {
+        return CustomAccountDto.from(accountRepository.findByAccountId(accountId).orElseThrow(NoMemberException::new));
+    }
+
+}
