--- src/main/java/com/sgyj/AccountServiceApplication.java
@@ -1,21 +1,0 @@
-package com.sgyj;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@ConfigurationPropertiesScan
-@SpringBootApplication
-@EnableDiscoveryClient
-@EnableJpaAuditing
-@EnableFeignClients
-public class AccountServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AccountServiceApplication.class, args);
-    }
-
-}
