--- src/main/java/com/sgyj/accountservice/infra/config/SecurityConfig.java
@@ -1,68 +1,0 @@
-package com.sgyj.accountservice.infra.config;
-
-import com.sgyj.accountservice.modules.account.service.AccountService;
-import com.sgyj.commonservice.properties.JwtProperties;
-import com.sgyj.commonservice.security.EntryPointHandler;
-import com.sgyj.commonservice.security.Jwt;
-import com.sgyj.commonservice.security.JwtAccessDeniedHandler;
-import com.sgyj.commonservice.security.JwtAuthenticationProvider;
-import com.sgyj.commonservice.security.JwtAuthenticationTokenFilter;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Slf4j
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final Jwt jwt;
-
-    private final JwtProperties jwtProperties;
-
-    private final JwtAccessDeniedHandler jwtAccessDeniedHandler;
-
-    private final EntryPointHandler unAuthorizedHandler;
-
-    @Bean
-    public JwtAuthenticationTokenFilter jwtAuthenticationTokenFilter() {
-        return new JwtAuthenticationTokenFilter(jwtProperties.getHeader(), jwt);
-    }
-
-
-    @Bean
-    public JwtAuthenticationProvider jwtAuthenticationProvider(AccountService accountService) {
-        return new JwtAuthenticationProviderImpl(accountService);
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http.httpBasic().disable().csrf().disable().cors().disable().exceptionHandling().accessDeniedHandler(jwtAccessDeniedHandler)
-            .authenticationEntryPoint(unAuthorizedHandler).and().headers().frameOptions().sameOrigin().and().sessionManagement()
-            .sessionCreationPolicy(SessionCreationPolicy.STATELESS).and().authorizeHttpRequests().anyRequest().permitAll().and()
-            .addFilterBefore(jwtAuthenticationTokenFilter(), UsernamePasswordAuthenticationFilter.class);
-        return http.build();
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
-        return authenticationConfiguration.getAuthenticationManager();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-
-}
