--- src/main/java/com/sgyj/accountservice/modules/account/dto/TokenDto.java
@@ -1,21 +1,0 @@
-package com.sgyj.accountservice.modules.account.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@Builder
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class TokenDto {
-
-    private String accessToken;
-
-    private String refreshToken;
-
-    public TokenDto(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-}
