--- src/main/java/com/sgyj/accountservice/modules/account/dto/kafka/Fields.java
@@ -1,14 +1,0 @@
-package com.sgyj.accountservice.modules.account.dto.kafka;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Fields {
-
-    private String type;
-    private boolean optional;
-    private String field;
-
-}
