--- src/main/java/com/sgyj/accountservice/modules/account/dto/kafka/Payload.java
@@ -1,11 +1,0 @@
-package com.sgyj.accountservice.modules.account.dto.kafka;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class Payload {
-
-
-}
