--- src/main/java/com/sgyj/accountservice/modules/account/dto/kafka/Schema.java
@@ -1,18 +1,0 @@
-package com.sgyj.accountservice.modules.account.dto.kafka;
-
-import java.util.List;
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class Schema {
-
-    private String type;
-    private List<Fields> fields;
-
-    private boolean optional;
-
-    private String name;
-
-}
