--- src/main/java/com/sgyj/accountservice/modules/account/enums/AccountStatus.java
@@ -1,11 +1,0 @@
-package com.sgyj.accountservice.modules.account.enums;
-
-public enum AccountStatus {
-
-    VERIFY_EMAIL,   // 이메일 인증 단계
-    NORMAL, // 정상
-    STOP,   // 중지
-    SLEEP,  // 휴면
-    ;
-
-}
