+++ src/test/java/com/gg/accountservice/AccountServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.gg.accountservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AccountServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
