--- src/test/java/com/sgyj/accountservice/AccountServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sgyj.accountservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AccountServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
